Fix public cache route exports
